Setup a basic framework for a new 'NDS File System' window
